pyproject.toml
[tool.poetry]
name = "cash"
version = "0.1.0"
description = "Cash exchange"
authors = [
    "Juan Pablo Garcia <juanpagfe@gmail..com>",
]
license = "GPLv2"
readme = "README.md"
homepage = "http://localhost"
repository = "http://localhost"

[tool.poetry.dependencies]
pyyaml = "^5.1"
python-http-client = "^3.2.4"
argparse = "^1.4.0"

[tool.poetry.scripts]
cash = 'cash.cli:run'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

